Some fixes for end-points & media timer

+++ worker/include/RTC/MediaTranslate/MediaTimer/MediaTimer.hpp
@@ -20,11 +20,11 @@ public:
     uint64_t RegisterTimer(const std::shared_ptr<MediaTimerCallback>& callback);
     void UnregisterTimer(uint64_t timerId);
     // time-out in milliseconds, previous invokes will discarded
-    void SetTimeout(uint64_t timerId, uint64_t timeoutMs);
+    void SetTimeout(uint64_t timerId, uint32_t timeoutMs);
     void Start(uint64_t timerId, bool singleshot);
     void Stop(uint64_t timerId);
     bool IsStarted(uint64_t timerId) const;
-    bool Singleshot(uint64_t afterMs, const std::shared_ptr<MediaTimerCallback>& callback);
+    bool Singleshot(uint32_t afterMs, const std::shared_ptr<MediaTimerCallback>& callback);
 private:
 	const std::shared_ptr<MediaTimerHandleFactory> _factory;
     const std::string _timerName;
