+++ worker/include/RTC/MediaTranslate/MediaTimer/MediaTimerHandle.hpp
@@ -11,18 +11,18 @@ class MediaTimerHandle
 {
 public:
     virtual ~MediaTimerHandle() = default;
-    uint64_t GetTimeout() const { return _timeoutMs.load(); }
-    void SetTimeout(uint64_t timeoutMs);
+    uint32_t GetTimeout() const { return _timeoutMs.load(); }
+    void SetTimeout(uint32_t timeoutMs);
     const auto& GetCallback() const { return _callback; }
     virtual void Start(bool singleshot) = 0;
     virtual void Stop() = 0;
     virtual bool IsStarted() const = 0;
 protected:
     MediaTimerHandle(const std::shared_ptr<MediaTimerCallback>& callback);
-    virtual void OnTimeoutChanged(uint64_t /*timeoutMs*/) {}
+    virtual void OnTimeoutChanged(uint32_t /*timeoutMs*/) {}
 private:
     const std::shared_ptr<MediaTimerCallback> _callback;
-    std::atomic<uint64_t> _timeoutMs = 0ULL;
+    std::atomic<uint32_t> _timeoutMs = 0ULL;
 };
 
 } // namespace RTC
